requirements.txt
arrow==1.3.0
binaryornot==0.4.4
blinker==1.8.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
create-app==0.6.2
Flask==3.0.3
Flask-JWT-Extended==4.6.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jinja2-time==0.2.0
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
requests==2.28.1
six==1.16.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.2
zipp==3.20.2
